Add workaround for area slicing issue
Choosing an sub area for data import in a scene objects like
EuropeCanary results in a wrong area slice due to wrong area
interpolation. If the lat lon values of a sub area are invalid
(e.g. in space) the slicing gets incorrect.
This commit will bypass this by calculating the slices directly
without interpolation for two areas with the same projection (geos)