Allow variations of image filename extensions
In a generic image filename extension, apart from `.png`, `.jpg`, and
`.tif`, also permit the long versions `.jpeg` and `.tiff`, as well as
uppercase versions.  Upper case filenames occur in the wild and a user
may be confused by a file `DEM.TIF` is rejected.  This commit does not
allow mixed case extensions as those are very unusual and it would make
a very long yaml file to allow them all.