Use NetCDF4FileHandler for FCI L1C FDHSI reader
For the FCI L1C FDHSI reader, use the NetCDF4FileHandler.  The previous
reader was keeping open every file three times per channel, leading to
many hundreds of open files.  This is not only unclean and inefficient,
but it may have been related to hard to predict segmentation faults we
were seeing.