Add corrected scale factors for FCI projection
The scale factors and add offsets for the MTG FCI test data as published
in 2019-09 are factually incorrect.  This probably causes the
differences between the pytroll and FCI geolocations.  Until this is
fixed from EUMETSATs side, hardcode corrected add_offset and
scale_factor values.

Similarly, there was a small error in the semi minor axis.

Also correct for an off-by-one error in the scale factors.