Adapt FCI filenames test
The FCI filenames matching test was buggy in multiple ways.  It was
passing six filenames and expecting four to match, even though five were
meeting the correct format.  This didn't fail in the past, because it
was using assertTrue rather than assertEquals.  Now that a previous
commit has fixed the assertion the test has started failing.  I don't
know why I ever expected 4 rather than 5 out of 6 filenames to pass,
this commit removes the one I was wrongly expecting to not pass from the
list of test filenames.