Add a feature for handling scheduled_time in ahi_hsd reader.
Added min and sec information to scheduled_time in ahi_hsd reader. The
scheduled_time had an accuracy of 10 min, but the Japan region (JPee)
and Region 3 (R3ff) are observed at a frequency of 150 sec.
Regions 4 (R4gg) and 5 (R5ii) of non-public data are observed at a
frequency of 30 sec. With this fix, scheduled_time is calculated
correctly for those observations.