Split DependencyTree from Node and DatatasetDict
In order to make future refactorings easier, DependencyTree is no longer a subclass of Node. Instead, the DependencyTree holds an instance of a Node as root.
Also, we make the DependencyTree rely on a new _DataIDContainer, which copies some functionality of the DatasetDict. However, this will allow for easier refactorings of the Scene and DependencyTree classes as they are separate.
Finally, this commit introduces tests for the DependencyTree class.