Apply valid_range in MiRS reader when present (#1655)
* Add reader_kwarg to omit limb correction on ATMS sensors.

* Fix docstring in mirs reader class

* add check for calling limb correction only when sensor is atms

* Fix docstring

* Add a kwarg test for limb_correction.

* Remove unused variables and add a noaa-20 test
Remove platform name from parameterize in last test for kwargs because
they are not used in that test.  Add a test for the NOAA-20 platform
when the limb_correction assertions are tested and platform name is
tested.

* Fold reader_kwarg test into basic_load

* Simply if/then statement, split parameterize for readability.

* Simplify assertion if/then statement for limb_correction.

* Remove extra line getting the name of the sensor

* Use valid range when present
Fixes a bug in which valid range was ignored.

* Check to confirm that valid range is no longer in attributes.

* Add a test to check valid range was applied correctly.
Change get_valid_range=>apply_valid_range and check for case of valid_range not None

* valid_range is inclusive so include both min/max in acceptable values
only read fake_data one time, not every time check_valid_range is called.