Change pmode approach in test
In the test for the pmode compositor, change the approach.  We don't
want to have to duplicate every composite for mode P and mode RGB
(currently), instead we would like that the user can pass an argument to
``load`` just like with different calibrations, resolutions, or
polarisations for single channels.  This should be doable via the
DataID...