fix: bad scan lines reafactor masking good scan lines
The refactor of mask_bad_scan_lines from np.choose to dataarray.where
inverted the logic of the line mask which lead to good scan lines being
masked.

Furthermore this fixes an undiscovered bug regarding the optional line
masking in the hrit reader introduced in in PR 1693 where the
corresponding reader quark was not recognized.