Change logic for ninjogeotiff gradient/axisintercept tags
In the ninjogeotiff writer, change the logic for deciding whether or
not to set the tags AxisIntercept and Gradient.  Previously, those were
always set, unless the image had mode RGB or RGBA.  With this change,
those tags are set only if the image has mode L and LA and PhysicUnit is
not set to N/A.

This commit also makes a small bugfix in a data fixture in the test
routine, that leads to a few reference values changing slightly.  This
reflects no change in functionality, but only between the fixtures and
the test functions.

Adapt documentation to reflect that NinJo 7 is now out and about.